Apply row styles while generating rows.
Before introducing nested tables, the length of the entire table could
be computer up-front and the cell styles could be applied to the entire
table once. With nested tables, the height of the table cannot be known
until after generating every row and nested table. The old code was
applying styles to an arbitrary square of cells, including top-level and
nested tables.